Refactored controllers in the modular style
