+++ config.ru
@@ -1,4 +1,4 @@
 require 'rubygems'
 require File.join(File.dirname(__FILE__), './app/chitter_app.rb')
 
-run Chitter
+run TheApp::Chitter
